requirements.txt
asgiref==3.5.2
cloudinary==1.30.0
Django==3.2.16
django-cloudinary-storage==0.3.0
django-filter==22.1
djangorestframework==3.14.0
Pillow==9.2.0
pytz==2022.4
sqlparse==0.4.3
